+++ component/admin/services/provider.php
@@ -12,9 +12,11 @@ use Joomla\CMS\Extension\MVCComponent;
 use Joomla\CMS\Extension\Service\Provider\ComponentDispatcherFactory;
 use Joomla\CMS\Extension\Service\Provider\MVCFactory;
 use Joomla\CMS\MVC\Factory\MVCFactoryInterface;
+use Joomla\Database\DatabaseInterface;
 use Joomla\DI\Container;
 use Joomla\DI\ServiceProviderInterface;
 use Xdecaro\Component\Decarodocuments\Administrator\Service\CoreIntegrationService;
+use Xdecaro\Component\Decarodocuments\Administrator\Service\RelationService;
 use Xdecaro\Component\Decarodocuments\Administrator\Service\StorageService;
 
 return new class () implements ServiceProviderInterface {
@@ -25,6 +27,10 @@ return new class () implements ServiceProviderInterface {
 
         $container->share(StorageService::class, static fn (): StorageService => new StorageService());
         $container->share(CoreIntegrationService::class, static fn (): CoreIntegrationService => new CoreIntegrationService());
+        $container->share(
+            RelationService::class,
+            static fn (Container $container): RelationService => new RelationService($container->get(DatabaseInterface::class))
+        );
 
         $container->set(
             ComponentInterface::class,